Show error if an exercise without assessmentmetadata is attempted to be rendered in an exam. Fixes #1301
